Refactored enableWhen
